Moved ContextualBalloon to the `panel/balloon` directory.
